backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
eventlet==0.41.0
Flask==3.1.3
flask-cors==6.0.2
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
wsproto==1.3.2
gunicorn==21.2.0
